Unify TextDocument printing for testing/debugging
