+++ compiler/test/dotty/tools/dotc/semanticdb/SemanticdbTests.scala
@@ -1,27 +1,25 @@
 package dotty.tools.dotc.semanticdb
 
 import scala.language.unsafeNulls
-
 import java.net.URLClassLoader
 import java.util.regex.Pattern
-import java.io.File
-import java.nio.file._
+import java.io.{ByteArrayOutputStream, File}
+import java.nio.file.*
 import java.nio.charset.StandardCharsets
 import java.util.stream.Collectors
 import java.util.Comparator
 import scala.util.control.NonFatal
 import scala.collection.mutable
-import scala.jdk.CollectionConverters._
-
+import scala.jdk.CollectionConverters.*
 import javax.tools.ToolProvider
-
-import org.junit.Assert._
+import org.junit.Assert.*
 import org.junit.Test
 import org.junit.experimental.categories.Category
-
 import dotty.BootstrappedOnlyTests
 import dotty.tools.dotc.Main
+import dotty.tools.dotc.semanticdb
 import dotty.tools.dotc.semanticdb.Scala3.given
+import dotty.tools.dotc.semanticdb.internal.SemanticdbOutputStream
 import dotty.tools.dotc.util.SourceFile
 
 @main def updateExpect =
@@ -153,41 +151,11 @@ class SemanticdbTests:
 end SemanticdbTests
 
 object SemanticdbTests:
-  /** Prettyprint a text document with symbol occurrences next to each resolved identifier.
-   *
-   * Useful for testing purposes to ensure that SymbolOccurrence values make sense and are correct.
-   * Example output (NOTE, slightly modified to avoid "unclosed comment" errors):
-   * {{{
-   *   class Example *example/Example#*  {
-   *     val a *example/Example#a.* : String *scala/Predef.String#* = "1"
-   *   }
-   * }}}
-   **/
   def printTextDocument(doc: TextDocument): String =
-    val symtab = doc.symbols.iterator.map(info => info.symbol -> info).toMap
-    val sb = StringBuilder(1000)
-    val sourceFile = SourceFile.virtual(doc.uri, doc.text)
-    var offset = 0
-    for occ <- doc.occurrences.sorted do
-      val range = occ.range.get
-      val end = math.max(
-        offset,
-        sourceFile.lineToOffset(range.endLine) + range.endCharacter
-      )
-      val isPrimaryConstructor =
-        symtab.get(occ.symbol).exists(_.isPrimary)
-      if !occ.symbol.isPackage && !isPrimaryConstructor then
-        assert(end <= doc.text.length,
-          s"doc is only ${doc.text.length} - offset=$offset, end=$end , symbol=${occ.symbol} in source ${sourceFile.name}")
-        sb.append(doc.text.substring(offset, end))
-        sb.append("/*")
-          .append(if (occ.role.isDefinition) "<-" else "->")
-          .append(occ.symbol.replace("/", "::"))
-          .append("*/")
-        offset = end
-    assert(offset <= doc.text.length, s"absurd offset = $offset when doc is length ${doc.text.length}")
-    sb.append(doc.text.substring(offset))
-    sb.toString
+    val byteStream = new ByteArrayOutputStream()
+    val out = SemanticdbOutputStream.newInstance(byteStream)
+    doc.writeTo(out)
+    out.flush()
+    semanticdb.textDocumentPrettyPrint(byteStream.toByteArray.nn)
   end printTextDocument
-
 end SemanticdbTests
