+++ presentation-compiler/test/dotty/tools/pc/utils/TestExtensions.scala
@@ -7,13 +7,3 @@ object TestExtensions:
     def getOffset(text: String): Int =
       val lines = text.linesWithSeparators.toList
       lines.take(pos.getLine).foldRight(0)(_.size + _) + pos.getCharacter
-
-  extension (range: dotty.tools.dotc.semanticdb.Range)
-    def getOffsets(text: String): (Int, Int) =
-      val lines = text.linesWithSeparators.toList
-      val startOffset = lines
-        .take(range.startLine)
-        .foldRight(0)(_.size + _) + range.startCharacter
-      val endOffset =
-        lines.take(range.endLine).foldRight(0)(_.size + _) + range.endCharacter
-      (startOffset, endOffset)
